requirements.txt
flask==2.0.1
python-dotenv==0.19.1
requests==2.26.0
pytrends==4.7.3
newsapi-python==0.2.6
flask-mail==0.9.1
beautifulsoup4==4.10.0
gunicorn==20.1.0
pytest==6.2.5
werkzeug==2.0.1
